Move data_distributor into app.
